requirements.txt
# Web Scraping
playwright==1.49.1
beautifulsoup4==4.12.3
lxml==5.3.0
aiohttp==3.11.11
aiofiles==24.1.0

# Document Processing
llama-parse==0.5.19
llama-index==0.12.10
python-dotenv==1.0.1

# Embeddings & Vector Store
openai==1.58.1
pinecone-client==5.0.1
tiktoken==0.8.0

# RAG
langchain==0.3.14
langchain-openai==0.3.0
langchain-pinecone==0.2.0

# API
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4

# Utilities
tqdm==4.67.1
pandas==2.2.3
httpx==0.28.1
tenacity==9.0.0
rich==13.9.4

# Web Server (Flask)
flask==3.0.0
gunicorn==21.2.0

# Reranking
sentence-transformers==2.2.2
rank_bm25==0.2.2

# Search
anthropic>=0.18.0
